Enable the C++ core guidelines checks in clang-tidy
In particular, I want the error when coroutines take byref arguments

+++ src/lib/utf8.cpp
@@ -89,7 +89,7 @@ std::string fold_utf8(std::string_view in) {
   ucasemap_utf8FoldCase(
     gFoldCaseMap,
     ret.data(),
-    ret.size(),
+    static_cast<int32_t>(ret.size()),
     in.data(),
     static_cast<int32_t>(in.size()),
     &error);
